feat: moved createIconikClient to helper/iconik.ts
